Move more extensions
