+++ Passepartout/Library/Sources/UILibrary/Extensions/ModuleBuilder+Previews.swift
@@ -1,8 +1,8 @@
 //
-//  ProfileManager+Editing.swift
+//  ModuleBuilder+Previews.swift
 //  Passepartout
 //
-//  Created by Davide De Rosa on 9/3/24.
+//  Created by Davide De Rosa on 8/19/24.
 //  Copyright (c) 2024 Davide De Rosa. All rights reserved.
 //
 //  https://github.com/passepartoutvpn
@@ -23,20 +23,25 @@
 //  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
 //
 
-import CommonLibrary
-import Foundation
 import PassepartoutKit
+import SwiftUI
 
-@MainActor
-extension ProfileManager {
-    func removeProfiles(at offsets: IndexSet) async {
-        let idsToRemove = headers
-            .enumerated()
-            .filter {
-                offsets.contains($0.offset)
-            }
-            .map(\.element.id)
+extension ModuleBuilder where Self: ModuleViewProviding {
 
-        await remove(withIds: idsToRemove)
+    @MainActor
+    public func preview(title: String = "") -> some View {
+        NavigationStack {
+            moduleView(with: ProfileEditor(modules: [self]), impl: nil)
+                .navigationTitle(title)
+        }
+        .withMockEnvironment()
+    }
+
+    @MainActor
+    public func preview<C: View>(with content: (ProfileEditor, Self) -> C) -> some View {
+        NavigationStack {
+            content(ProfileEditor(modules: [self]), self)
+        }
+        .withMockEnvironment()
     }
 }
